requirements.txt
aiohappyeyeballs==2.6.1
aiobotocore==2.23.0
aiohttp==3.12.13
aioitertools==0.12.0
aiosignal==1.3.2
annotated-types==0.7.0
attrs==25.3.0
botocore==1.38.27
bounded-pool-executor==0.0.3
cdsapi==0.7.6
certifi==2025.6.15
charset-normalizer==3.4.2
contourpy==1.3.2
cycler==0.12.1
earthaccess==0.14.0
<<<<<<< Updated upstream
ecmwf-datastores-client==0.2.0
fonttools==4.58.4
frozenlist==1.7.0
fsspec==2025.5.1
idna==3.10
importlib_resources==6.5.2
jmespath==1.0.1
kiwisolver==1.4.8
multidict==6.5.0
multimethod==2.0
multiurl==0.3.5
packaging==25.0
pandas==2.3.0
pillow==11.2.1
pip==25.1.1
pqdm==0.2.0
propcache==0.3.2
pydantic==2.11.7
pydantic-settings==2.9.1
pydantic_core==2.33.2
pyparsing==3.2.3
python-cmr==0.13.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
=======
earthengine_api==1.5.22
pydantic==2.11.7
python-dotenv==1.1.1
PyYAML==6.0.2
>>>>>>> Stashed changes
PyYAML==6.0.2
requests==2.32.4
s3fs==2025.5.1
setuptools==80.9.0
six==1.17.0
tinynetrc==1.3.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
wheel==0.45.1
wrapt==1.17.2
yarl==1.20.1